requirements.txt
appnope==0.1.4
arch==8.0.0
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
frozendict==2.4.7
fsspec==2026.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
learn==1.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
multitasking==0.0.12
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pandas==3.0.0
parso==0.8.5
patsy==1.0.2
peewee==3.19.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
sci-kit==0.1.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.5.0
websockets==16.0
yfinance==1.1.0
